Conflicts:
	CMakeLists.txt
	main/include/eudaq/Event.hh
	main/include/eudaq/FileWriter.hh
	main/include/eudaq/Serializer.hh

+++ main/include/eudaq/DataConverterPlugin.hh
@@ -17,7 +17,7 @@ using namespace UTIL;
 #endif
 #include "TLUEvent.hh"
 
-#define NOTIMESTAMPSET (unsigned long long)-1
+#define NOTIMESTAMPSET (uint64_t)-1
 #define NOTIMEDURATIONSET 0
