+++ main/lib/src/FileWriterDetectorEvents.cc
@@ -13,13 +13,13 @@ namespace eudaq {
       FileWriterTextdetector(const std::string &);
       virtual void StartRun(unsigned);
       virtual void WriteEvent(const DetectorEvent &);
-      virtual unsigned long long FileBytes() const;
+      virtual uint64_t FileBytes() const;
       virtual ~FileWriterTextdetector();
     private:
     //  std::FILE * m_file;
 	  std::ofstream *m_out;
 	  bool firstEvent;
-	  unsigned long long int DUT_start_time;
+	  uint64_t DUT_start_time;
   };
 
   namespace {
@@ -95,6 +95,6 @@ namespace eudaq {
     }
   }
 
-  unsigned long long FileWriterTextdetector::FileBytes() const { return m_out->tellp(); }
+  uint64_t FileWriterTextdetector::FileBytes() const { return m_out->tellp(); }
 
 }
