+++ main/lib/src/Time.cc
@@ -5,16 +5,12 @@
 #include <iostream>
 
 #if EUDAQ_PLATFORM_IS(WIN32)
-# define WIN32_LEAN_AND_MEAN
-//# include <afxwin.h> 
-# include <Windows.h>
-# include <time.h>
-# define EPOCHFILETIME (116444736000000000LL)
+#define WIN32_LEAN_AND_MEAN
+#include <time.h>
+#define EPOCHFILETIME (116444736000000000LL)
 #endif
 
 namespace eudaq {
-
-
 
 
 #if EUDAQ_PLATFORM_IS(WIN32)
@@ -33,8 +29,8 @@ namespace eudaq {
     t /= 10;                /* In microseconds */
 
     timeval tv;
-    tv.tv_sec  = (long)(t / 1000000);
-    tv.tv_usec = (long)(t % 1000000);
+    tv.tv_sec  = static_cast<int32_t>(t / 1000000);
+    tv.tv_usec = static_cast<int32_t>(t % 1000000);
 
     return tv;
   }
@@ -61,7 +57,7 @@ namespace eudaq {
     time.tm_hour = hour;
     time.tm_min = minute;
     time.tm_sec = sec + usec / 1000000;
-    tv_sec = static_cast<long>(mktime(&time));
+    tv_sec = static_cast<int32_t>(mktime(&time));
     tv_usec = usec % 1000000;
   }
